Base events on new NavigatedTelemetryEvent
